requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --strip-extras requirements/base.in
#
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
click==8.1.7
    # via
    #   -r requirements/base.in
    #   uvicorn
colorama==0.4.6
    # via click
dnspython==2.4.2
    # via pymongo
exceptiongroup==1.1.3
    # via anyio
fastapi==0.104.1
    # via -r requirements/base.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
mongoengine==0.27.0
    # via -r requirements/base.in
pydantic==2.4.2
    # via
    #   -r requirements/base.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-settings==2.0.3
    # via -r requirements/base.in
pymongo==4.5.0
    # via mongoengine
python-dotenv==1.0.0
    # via pydantic-settings
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   uvicorn
uvicorn==0.23.2
    # via -r requirements/base.in
